Extract geometry helpers
