+++ frontend/src/spells.js
@@ -19,7 +19,8 @@ export function fireballStats(level) {
   return { damage, radius, pierce };
 }
 
-import { circleRectColliding, isColliding } from "./game_logic.js";
+import { circleRectColliding } from "./game_logic.js";
+import { isColliding } from "./utils/geometry.js";
 import { damageWall } from "./walls.js";
 export function createFireball(x, y, direction, level = 1, damageMult = 1) {
   const len = Math.hypot(direction.x, direction.y);
