+++ frontend/src/utils/geometry.js
@@ -1,0 +1,20 @@
+// Generic 2D geometry helpers used across the game
+
+export function moveTowards(entity, target, speed) {
+  const dx = target.x - entity.x;
+  const dy = target.y - entity.y;
+  const dist = Math.hypot(dx, dy);
+  if (dist === 0) return;
+  entity.x += (dx / dist) * speed;
+  entity.y += (dy / dist) * speed;
+  if (entity.facing) {
+    entity.facing.x = dx / dist;
+    entity.facing.y = dy / dist;
+  }
+}
+
+export function isColliding(a, b, radius) {
+  const dx = a.x - b.x;
+  const dy = a.y - b.y;
+  return Math.hypot(dx, dy) < radius * 2;
+}
